+++ src/components/films-lists.js
@@ -1,0 +1,38 @@
+const createFilmsListTemplate = () => {
+  return (
+    `<ul class="sort">
+      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
+      <li><a href="#" class="sort__button">Sort by date</a></li>
+      <li><a href="#" class="sort__button">Sort by rating</a></li>
+    </ul>
+
+    <section class="films">
+      <section class="films-list">
+        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+        <div class="films-list__container"></div>
+      </section>
+    </section>`
+  );
+};
+
+const createTopRatedListTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Top rated</h2>
+
+      <div class="films-list__container"></div>
+    </section>`
+  );
+};
+
+const createMostCommentedListTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Most commented</h2>
+
+      <div class="films-list__container"></div>
+    </section>`
+  );
+};
+
+export {createFilmsListTemplate, createMostCommentedListTemplate, createTopRatedListTemplate};
